src/virtual_networking.py
```python
import random
from src.chatbot import ChatBot
from src.breakout_rooms import BreakoutRoom

class VirtualNetworking:
    def __init__(self, userData, eventData):
        self.userData = userData
        self.eventData = eventData
        self.chatBot = ChatBot()
        self.breakoutRoom = BreakoutRoom()

    def startLiveChat(self, userId, eventId):
        user = self.userData[userId]
        event = self.eventData[eventId]
        self.chatBot.startChat(user, event)

    def createBreakoutRoom(self, eventId):
        event = self.eventData[eventId]
        self.breakoutRoom.createRoom(event)

    def network(self, userId, eventId):
        user = self.userData[userId]
        event = self.eventData[eventId]
        attendees = event['attendees']
        potentialConnections = [attendee for attendee in attendees if attendee != userId]
        connection = random.choice(potentialConnections)
        return connection
```