tests/test_virtual_networking.py
```python
import unittest
from src import virtual_networking

class TestVirtualNetworking(unittest.TestCase):

    def setUp(self):
        self.networking = virtual_networking.VirtualNetworking()
        self.user_data = {"userId": 1, "userName": "Test User", "userInterests": ["AI", "Networking"]}
        self.event_data = {"eventId": 1, "eventName": "AI Conference", "eventTopics": ["AI", "Networking"]}

    def test_start_live_chat(self):
        result = self.networking.start_live_chat(self.user_data)
        self.assertEqual(result, "Live chat started for user: Test User")

    def test_create_breakout_room(self):
        result = self.networking.create_breakout_room(self.event_data)
        self.assertEqual(result, "Breakout room created for event: AI Conference")

    def test_network(self):
        result = self.networking.network(self.user_data, self.event_data)
        self.assertEqual(result, "Networking started for user: Test User at event: AI Conference")

if __name__ == '__main__':
    unittest.main()
```